feat: support config.yaml

+++ preprocess.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
-wget -O /root/.config/clash/config.yaml "http://subconverter:25500/sub?proxy_provider=true&target=clash&url=${RAW_SUB_URL}"
+
+if [ ! -f /root/.config/clash/config.yaml ]; then
+    wget -O /root/.config/clash/config.yaml "http://subconverter:25500/sub?proxy_provider=true&target=clash&url=${RAW_SUB_URL}"
+fi
+
 # Switch to the container command
 exec "/clash"
